chore(Overlay): moves modal and sidesheet common functions in utils
