+++ core/utils/overlayHelper.ts
@@ -1,0 +1,34 @@
+export const getWrapperElement = (): Element => {
+  let element = document.querySelector('.Overlay-wrapper');
+  if (element === null) {
+    element = document.createElement('div');
+    element.classList.add('Overlay-wrapper');
+    document.body.appendChild(element);
+  }
+  return element;
+};
+
+interface elementData {
+  element: Element;
+  containerClassName: string;
+  elementRef: React.RefObject<HTMLDivElement>;
+}
+
+export const getUpdatedZIndex = (ele: elementData): number | undefined => {
+  const { containerClassName, elementRef, element } = ele;
+
+  if (element === null) return;
+
+  const elements = element.querySelectorAll(containerClassName);
+  if (elements.length < 1) return;
+
+  const siblings = Array.from(elements).filter(el => el !== elementRef.current);
+  let zIndex = -1;
+
+  siblings.forEach(element => {
+    const prevZIndex = parseInt(window.getComputedStyle(element).zIndex || '0', 10);
+    zIndex = Math.max(zIndex, prevZIndex + 10);
+  });
+
+  return zIndex > 0 ? zIndex : undefined;
+};
